# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SH3 API Services",
        "description": "API developed for SH3 backend",
        "contact": {
            "name": "Gabriel Meireles",
            "email": "gabriel.meireles@sh3.com.br"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/postTicket": {
            "post": {
                "tags": [
                    "Ticket"
                ],
                "summary": "Create a new ticket",
                "description": "Create a new ticket with the provided details",
                "operationId": "702dd649f84888ba0b63b472f31b741f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "client",
                                    "occupation_area"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Ticket name",
                                        "type": "string"
                                    },
                                    "client": {
                                        "description": "Client name or identifier",
                                        "type": "string"
                                    },
                                    "occupation_area": {
                                        "description": "Occupation area of the ticket",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ticket created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Ticket Name"
                                            },
                                            "client": {
                                                "type": "string",
                                                "example": "Client Name"
                                            },
                                            "occupation_area": {
                                                "type": "string",
                                                "example": "Occupation Area"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error or missing required fields",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Error message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/putTicket": {
            "put": {
                "tags": [
                    "Ticket"
                ],
                "summary": "Update a ticket",
                "description": "Update the details of an existing ticket",
                "operationId": "3c8d53822d6f01f4e66566196eb026d3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id",
                                    "name",
                                    "client",
                                    "occupation_area"
                                ],
                                "properties": {
                                    "id": {
                                        "description": "ID of the ticket to be updated",
                                        "type": "integer"
                                    },
                                    "name": {
                                        "description": "Updated ticket name",
                                        "type": "string"
                                    },
                                    "client": {
                                        "description": "Updated client name or identifier",
                                        "type": "string"
                                    },
                                    "occupation_area": {
                                        "description": "Updated occupation area of the ticket",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ticket updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Ticket Name"
                                            },
                                            "client": {
                                                "type": "string",
                                                "example": "Client Name"
                                            },
                                            "occupation_area": {
                                                "type": "string",
                                                "example": "Occupation Area"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ticket not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Error message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error or missing required fields",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Error message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/getTickets": {
            "get": {
                "tags": [
                    "Ticket"
                ],
                "summary": "Get all tickets",
                "description": "Retrieve a list of all tickets",
                "operationId": "31ed28d3dc400b5fbec6c654ef219d4b",
                "responses": {
                    "200": {
                        "description": "List of all tickets",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Ticket Name"
                                            },
                                            "client": {
                                                "type": "string",
                                                "example": "Client Name"
                                            },
                                            "occupation_area": {
                                                "type": "string",
                                                "example": "Occupation Area"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tickets not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Error message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/getTicket": {
            "get": {
                "tags": [
                    "Ticket"
                ],
                "summary": "Get ticket by ID",
                "description": "Retrieve a ticket by its ID",
                "operationId": "89d76b40f80767eccb9d8c16c71e3e11",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the ticket",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ticket found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Ticket Name"
                                        },
                                        "client": {
                                            "type": "string",
                                            "example": "Client Name"
                                        },
                                        "occupation_area": {
                                            "type": "string",
                                            "example": "Occupation Area"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ticket not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Error message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/deleteTicket": {
            "delete": {
                "tags": [
                    "Ticket"
                ],
                "summary": "Delete ticket by ID",
                "description": "Delete a ticket by its ID",
                "operationId": "be494f0a27020892a47bc68757cc7d80",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the ticket",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Ticket deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ticket deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ticket not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ticket not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "scheme": "bearer",
                "description": "Enter token in format (Bearer Token)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}